[feat]발행쿠폰 등록화면 수정(유효기간 제한없음 추가) / 상세보기 매퍼 sql문 수정, 카테고리 표기용 뷰테이블 생성, 개별쿠폰 발행 진행중

+++ src/main/java/util/NumberFormatting.java
@@ -19,7 +19,7 @@ public class NumberFormatting {
     }
 	
 	//금액 숫자 -> 금액단위로 포매팅
-	public String formatMoney(long amount) {
+	public String formatMoney(Integer amount) {
 		if (amount < 1_000) {
 	        return amount + "원";
